requirements.txt
# Research Vision - Production Dependencies
# Python 3.8+

# Core dependencies
opencv-python>=4.8.0
numpy>=1.24.0
PyMuPDF>=1.23.0
Pillow>=10.0.0

# Web interface
streamlit>=1.28.0
